requirements.txt
PyPDF2>=3.0.0
pandas>=1.5.0
matplotlib>=3.6.0
pillow>=9.0.0 